Remove elastic_doc_manager (now packaged separately)
