+++ .travis.yml
@@ -6,19 +6,14 @@ python:
   - 3.4
   - 3.5
 
-services:
-  - elasticsearch
-
 install:
   - pip install pymongo==2.9
   - pip install mongo-orchestration==0.3.1
   - curl -O http://archive.apache.org/dist/lucene/solr/4.9.1/solr-4.9.1.tgz
   - tar -zxf solr-4.9.1.tgz
   - cp mongo_connector/doc_managers/schema.xml solr-4.9.1/example/solr/collection1/conf/
-  - sudo /usr/share/elasticsearch/bin/plugin install elasticsearch/elasticsearch-mapper-attachments/2.4.3
 
 before_script:
-  - sudo service elasticsearch restart
   - cd solr-4.9.1/example/ && java -Djetty.port=8983 -Dsolr.solr.home=solr -jar start.jar > /dev/null 2>&1 &
   - mongo-orchestration start
