+++ tests/test_connector_sharded.py
@@ -1,11 +1,26 @@
+# Copyright 2013-2016 MongoDB, Inc.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+# http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
 import os
 
 from mongo_connector.connector import Connector
 from mongo_connector.doc_managers.doc_manager_simulator import DocManager
-
-from tests import unittest, SkipTest, db_user, db_password
-from tests.setup_cluster import ShardedCluster
-from tests.util import assert_soon
+from mongo_connector.test_utils import (ShardedCluster,
+                                        db_user,
+                                        db_password,
+                                        assert_soon)
+from tests import unittest, SkipTest
 
 
 class TestConnectorSharded(unittest.TestCase):
