+++ tests/test_mongo_doc_manager.py
@@ -21,8 +21,8 @@ sys.path[0:0] = [""]
 
 from mongo_connector.command_helper import CommandHelper
 from mongo_connector.doc_managers.mongo_doc_manager import DocManager
-from tests import unittest, TESTARGS
-from tests.test_gridfs_file import MockGridFSFile
+from mongo_connector.test_utils import TESTARGS, MockGridFSFile
+from tests import unittest
 from tests.test_mongo import MongoTestCase
