+++ tests/test_rollbacks.py
@@ -1,3 +1,17 @@
+# Copyright 2013-2016 MongoDB, Inc.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+# http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
 """Test Mongo Connector's behavior when its source MongoDB system is
 experiencing a rollback.
 
@@ -12,14 +26,15 @@ from pymongo.write_concern import WriteConcern
 
 sys.path[0:0] = [""]
 
-from mongo_connector.util import retry_until_ok
-from mongo_connector.locking_dict import LockingDict
 from mongo_connector.doc_managers.doc_manager_simulator import DocManager
+from mongo_connector.locking_dict import LockingDict
 from mongo_connector.oplog_manager import OplogThread
-
-from tests import unittest, STRESS_COUNT
-from tests.util import assert_soon, close_client
-from tests.setup_cluster import ReplicaSet
+from mongo_connector.test_utils import (ReplicaSet,
+                                        STRESS_COUNT,
+                                        assert_soon,
+                                        close_client)
+from mongo_connector.util import retry_until_ok
+from tests import unittest
 
 
 class TestRollbacks(unittest.TestCase):
